WIP split tx_generator to multiple files
